pyproject.toml
[tool.poetry]
name = "ell-test"
version = "0.1.0"
description = ""
authors = ["Olaf Geibig <GEO5BE4@bosch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ell-ai = "^0.0.13"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
openai = "^1.47.0"
anthropic = "^0.34.2"
instructor = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
